Helpful error message if serve-files requirements not available
