+++ extra_data/utils.py
@@ -9,7 +9,6 @@ program. If not, see <https://opensource.org/licenses/BSD-3-Clause>
 """
 
 import os
-from socket import AF_INET
 from warnings import warn
 
 import h5py
@@ -156,20 +155,6 @@ def hdf5_to_cbf(in_h5file, cbf_filename, index, header=None):
     cbf_out = numpy_to_cbf(images, index=index)
     cbf_out.write(cbf_filename)
     print("Convert {} index {} to {}".format(in_h5file, index, cbf_filename))
-
-
-def find_infiniband_ip():
-    """Find the first infiniband IP address
-
-    :returns: str
-        IP of the first infiniband interface if it exists else '*'
-    """
-    from psutil import net_if_addrs
-    addrs = net_if_addrs()
-    for addr in addrs.get('ib0', ()):
-        if addr.family is AF_INET:
-            return addr.address
-    return '*'
 
 
 def available_cpu_cores():
